Make the quote and order stuff separate from the wtg/catalog module

+++ src/Migrations/2017_03_28_130100_create_order_items_table.php
@@ -17,7 +17,7 @@ class CreateOrderItemsTable extends Migration
             $table->uuid('order_id');
             $table->string('sku');
             $table->string('name');
-            $table->double('quantity');
+            $table->float('quantity');
         });
     }
